\name{ZILGM-package}
\alias{ZILGM-package}
\docType{package}
\title{
A R Package to estimate Local Markov Network for Count data with zero-inflation and overdispersion.
}
\description{
Undirected graphical models to estimate Local Markov Network for Count data with zero-inflation and overdispersion.
}

\details{
\tabular{rl}{
Package: \tab ZILGM\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2021-02-22\cr
License: \tab GPL-2\cr
}
}

\author{
Beomjin, Park.

Maintainer : Beomjin, Park. <bbbeomjin@gmail.com>
}

\references{Choi, H., Gim, J., Won, S., Kim, Y. J., Kwon, S., and Park, C. (2017). Network analysis for count data with excess zeros. \emph{BMC genetics}, \bold{18}(1), 93. \cr
			Park, B., Choi, H., Park, C. (2021). Negative binomial graphical model with excess zeros.}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\examples{
library(devtools)
install_github(bbeomjin/ZILGM)
library(ZILGM)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
