\name{classic3}
\alias{classic3}
\title{
Classic3 dataset
}
\description{
The classic3 dataset consists of CISI, MED, and CRAN document collections. CISI, MED, and CRAN contain 1460, 1032, and 1398 abstracts from information science, medical, and aerospace engineering journals, respectively.
}

\usage{
data(classic3)
}

\format{
  A data frame with 3890 rows and 5898 variables.
}

\details{
The raw dataset of the classic3 and classic4 dataset can be found at \url{http://www.dataminingresearch.com/index.php/2010/09/classic3-classic4-datasets}.
}

\references{Classic3 dataset, 2009: Classic3 dataset. Retrieved November 29, 2009 from World Wide Web, available at \url{http://www.dataminingresearch.com/index.php/2010/09/classic3-classic4-datasets}}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\examples{
require(ZILGM)
data(classic3)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
