\name{find_lammax}
\alias{find_lammax}
\title{
Compute the maximum \code{lambda}
}
\description{
Compute the maximum \code{lambda} from Karush-Kuhn-Tucker condition
}
\usage{find_lammax(X)}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{X}{
  A \emph{n} x \emph{p} data matrix.
}
}
\details{
Zero-inflated local graphical model
}
\value{
  a numeric value
}
\references{Choi, H., Gim, J., Won, S., Kim, Y. J., Kwon, S., and Park, C. (2017). Network analysis for count data with excess zeros. \emph{BMC genetics}, \bold{18}(1), 93.}
\author{
Park Beomjin, 
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\examples{
Not yet
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
