\name{find_lammax}
\alias{find_lammax}
\title{
Compute the maximum lambda
}
\description{
Compute the maximum lambda
}
\usage{find_lammax(X)}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{X}{
  A $n \times p$ data matrix.
}
}
\details{
Zero-inflated local graphical model
}
\value{
  An S3 object with the following slots
  \item{lammax}{a maximum lambda from data matrix}
}
\references{Network analysis for excess zero}
\author{
Park Beomjin, 
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\examples{
Not yet
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
